Adds "Use Internal Action Filter" Action Filter
With PR #10, we've had to swap out our calls to `contentResource.getScaffold` resource, as that doesn't work with a Guid ID.
However in doing that we actually lose an Umbraco goodie, namely that the api data is passed through the `OutgoingEditorModelEvent` action filter

https://github.com/umbraco/Umbraco-CMS/blob/release-7.4.0/src/Umbraco.Web/Editors/ContentController.cs#L144

Unfortunately the `OutgoingEditorModelEvent` attribute is marked as internal, so we can't use it directly.

https://github.com/umbraco/Umbraco-CMS/blob/release-7.4.0/src/Umbraco.Web/WebApi/Filters/OutgoingEditorModelEventAttribute.cs

I initially added an action filter that used reflection to only call the `OutgoingEditorModelEventAttribute.OnActionExecuted` method.
But then abstracted it to handle any other action filters that are marked as internal.