Self-heals "single-item mode" items when upgrading to v1.1.0
When upgrading to v1.1.0 (currently in alpha), the JS code will attempt self-heal the data-type and property-data values,
(note this is on the client-side and data-types would need to be re-saved for this to take effect).

However for any empty "single-item mode" items, the content-type's scaffold is requested on-load.
But the JS code is only looking to use the `icContentTypeGuid` value, whereas the property's data-type only has the `icContentTypeAlias` value available.

One fix is to re-save the property's data-type, job done - however if a user forgets to do this, then they will see an error in the Content section.

This patch checks if the `icContentTypeGuid` value is available, otherwise falls back on the old `icContentTypeAlias` value.